requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-allauth==0.44.0
django-appconf==1.0.4
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-phonenumber-field==5.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docopt==0.6.2
filelock==3.0.12
idna==2.10
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
markdown2==2.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.12.22
pilkit==2.0
Pillow==8.2.0
pipreqs==0.4.10
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
yarg==0.1.9